requirements.txt
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==17.4.0
Automat==0.6.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==6.7
cloud-init==20.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
incremental==16.10.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
lxml==4.6.2
lyricsgenius==2.0.2
MarkupSafe==1.0
matplotlib==3.3.3
MechanicalSoup==0.12.0
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.4
oauthlib==2.0.6
packaging==20.4
PAM==0.4.2
pandas==0.22.0
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.2
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-Levenshtein==0.12.0
pytz==2020.5
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
six==1.11.0
soupsieve==2.1
ssh-import-id==5.7
systemd-python==234
terminado==0.9.1
testpath==0.4.4
tornado==6.1
tqdm==4.19.5
traitlets==4.3.3
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
zope.interface==4.3.2
